Print cluster diagnostics
