+++ continuous_integration/simple_test.sh
@@ -3,4 +3,4 @@ source activate test
 conda install -c conda-forge -y lxml py4j
 cd /knit
 python setup.py install mvn
-py.test -vv
+py.test -vv -s
