+++ knit/tests/test_core.py
